% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{CreateGrid}
\alias{CreateGrid}
\title{Creates a grid in cartesian space for interpolation function.}
\usage{
CreateGrid(sensor.space, grain = 40L)
}
\arguments{
\item{sensor.space}{A matrix. This matrix should contain the x,y locations of
  the sensors.}

\item{grain}{A scalar. This represents that spacing between the points in the
  grid. Defaults to 40, to create a relatively coarse surface. Higher values
  will create a more fine-grained, and therefore smoother-looking surface.}
}
\value{
This function returns a data.frame with the cartesian coordinates of
  the bounding rectangle for the interpolation function.
}
\description{
Creates a grid in cartesian space for interpolation function.
}
\details{
This function takes a matrix of cartesian coordinates of EEG/MEG
  sensors and creates a regularly-spaced grid over the sensor space bounding
  rectangle. A denser grid will give a very smooth-looking interpolated
  surface, but will take longer to compute.
}

