% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{CreateInterpolationSurface}
\alias{CreateInterpolationSurface}
\title{Creates an interpolation surface in cartesian space for EEG/MEG
  sensors.}
\usage{
CreateInterpolationSurface(x, y = NULL)
}
\arguments{
\item{x}{A matrix or a vector. If x is a matrix, it needs to have two columns
  containig the (x, y) locations of the sensors, in which case y is not
  required. If it is a vector, it should contain the x-coordinates of the
  sensor.}

\item{y}{A vector containing the y-locations of the sensors. Defaults to
  NULL, but will throw an error if x does not contain all the required
  information.}
}
\value{
This function returns a data.frame with the cartesian coordinates of
  the surface for interpolation.
}
\description{
Creates an interpolation surface in cartesian space for EEG/MEG
  sensors.
}

