% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{FindBoundingRectangle}
\alias{FindBoundingRectangle}
\title{Finds the bounding rectangle for the sensors in cartesian space.}
\usage{
FindBoundingRectangle(sensor.space, inflation = 1.05)
}
\arguments{
\item{sensor.space}{A matrix. This matrix should contain the x,y locations of
  the sensors.}

\item{inflation}{A scalar. This represents a small fudge factor to make sure
  that the bounding rectangle will include all the sensors, in case there are
  sensors that would fall exactly in the rectangle line. Defaults to 5% of
  the rectangle dimensions.}
}
\value{
This function returns a data.frame with the cartesian coordinates of
  the bounding rectangle for the interpolation function.
}
\description{
Finds the bounding rectangle for the sensors in cartesian space.
}

