% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{FormatPosition}
\alias{FormatPosition}
\title{Formats the x and y locations of EEG/MEG sensors for interpolation
  function.}
\usage{
FormatPosition(x, y = NULL)
}
\arguments{
\item{x}{A matrix or a vector. If x is a matrix, it needs to have two columns
  containing the (x, y) locations of the sensors, in which case y is not
  required. If it is a vector, it should contain the x-coordinates of the
  sensor.}

\item{y}{A vector containing the y-locations of the sensors. Defaults to
  NULL, but will throw an error if x does not contain all the required
  information.}
}
\value{
This function returns a data.frame with two fields (x, y), properly
  formatted for the interpolation function.
}
\description{
Formats the x and y locations of EEG/MEG sensors for
  interpolation function.
}

