% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{IsInConvexHull}
\alias{IsInConvexHull}
\title{Tests whether a point is inside the convex hull created by the EEG/MEG
  sensors.}
\usage{
IsInConvexHull(interpolation.surface, electrode.positions)
}
\arguments{
\item{interpolation.surface}{A matrix containig the (x, y) locations of the
  surface to be interpolated.}

\item{electrode.positions}{A matrix containing the (x, y) locations of the
  EEG/MEG sensors.}
}
\value{
This function returns TRUE or FALSE depending on whether a point is
  inside the convex hull.
}
\description{
Tests whether a point is inside the convex hull created by the
  EEG/MEG sensors.
}
\details{
Calls function in.convex.hull from package tripack.
}

