% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{MakeSymmetric}
\alias{MakeSymmetric}
\title{Make a symmetric range for color bar}
\usage{
MakeSymmetric(z.values)
}
\arguments{
\item{z.values}{A vector of potentials.}
}
\value{
vector of length 2 with symmetric range.
}
\description{
Make a symmetric range for color bar
}

