% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{MakeSymmetricCuts}
\alias{MakeSymmetricCuts}
\title{Create range for asymmetric color bar}
\usage{
MakeSymmetricCuts(z.values, steps)
}
\arguments{
\item{z.values}{A vector of potentials.}

\item{steps}{A scalar. Number of steps in color bar.}
}
\value{
vector of length 'steps' for color bar.
}
\description{
Create labels for symmetric color bar
}

