% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{MakeSymmetricLabels}
\alias{MakeSymmetricLabels}
\title{Create labels for symmetric color bar}
\usage{
MakeSymmetricLabels(z.values, multiplication.factor = 10, unit = NULL)
}
\arguments{
\item{z.values}{A vector of potentials.}

\item{multiplication.factor}{Correction factor for graphic display of units
  of measurement (generally microvolts or femtotesla).}

\item{unit}{A string. Can be either 'mV' (for microvolts, usually for EEG
  data), 'fT' (for femtotesla, usually for MEG data) or 'mVcm2' (for
  microvolts over squared centimeters, usually for CSD transformed EEG data).}
}
\value{
vector of length 3 with symmetric range around 0.
}
\description{
Create labels for symmetric color bar
}

