% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SensorSpaceInterpolation.R
\name{SimplifyTopographyList}
\alias{SimplifyTopographyList}
\title{Simplify a list to save space}
\usage{
SimplifyTopographyList(topography.list)
}
\arguments{
\item{topography.list}{A list. Natural output of function
  InterpolatePotentials}
}
\value{
List with interpolated potentials for each time slice, simplified.
}
\description{
Simplify a list to save space.
}
\details{
The output of function InterpolatePotentials is very redundant.
  This function simplifies it by removing the x, y fields from every item on
  the list, since they are always the same.
}

